Add Transaction Functional

+++ app/src/main/java/com/example/movilesapp/model/entities/User.kt
@@ -3,7 +3,7 @@ package com.example.movilesapp.model.entities
 data class User(
     val userId: String,
     val name: String,
-    val phone: Int,
+    val phone: Long,
     val email: String,
-    val balance: Int
+    val balance: Double
 )
